reorg: move code related to static compilation to precompile.c
also fix compile=all for jb/subtype
